requirements.txt
Click==7.0
Flask==1.1.1
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
passlib==1.7.2
pkg-resources==0.0.0
Werkzeug==0.16.0
